requirements.txt
datasets>=3.0.0
haystack_ai>=2.5.1
sentence_transformers>=3.1.0